metrics. Also, functions in separate file
